center cluster数据同步增加重试并能保证数据幂等

+++ dao-cloud-center/src/main/java/com/junmo/center/core/handler/SyncClusterInformationRequestHandler.java
@@ -4,26 +4,34 @@ import com.junmo.center.core.ConfigCenterManager;
 import com.junmo.center.core.RegisterCenterManager;
 import com.junmo.core.model.ClusterSyncDataModel;
 import com.junmo.core.model.RegisterProviderModel;
+import com.junmo.core.util.ExpireHashMap;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
+
+import java.util.concurrent.TimeUnit;
 
 /**
  * @author: sucf
  * @date: 2023/6/24 00:00
- * @description:
+ * @description: 请求处理cluster数据同步
  */
-public class SyncClusterInformationHandler extends SimpleChannelInboundHandler<ClusterSyncDataModel> {
+public class SyncClusterInformationRequestHandler extends SimpleChannelInboundHandler<ClusterSyncDataModel> {
 
     private ConfigCenterManager configCenterManager;
 
-    public SyncClusterInformationHandler(ConfigCenterManager configCenterManager) {
+    private ExpireHashMap<Long> expireHashMap;
+
+    public SyncClusterInformationRequestHandler(ConfigCenterManager configCenterManager) {
+        expireHashMap = new ExpireHashMap(1000, 1, TimeUnit.HOURS);
         this.configCenterManager = configCenterManager;
     }
 
     @Override
     protected void channelRead0(ChannelHandlerContext ctx, ClusterSyncDataModel clusterSyncDataModel) {
         if (clusterSyncDataModel.getType() == -2) {
-            configCenterManager.delete(clusterSyncDataModel.getProxyConfigModel());
+            if (!expireHashMap.exists(clusterSyncDataModel.getSequenceId())) {
+                configCenterManager.delete(clusterSyncDataModel.getProxyConfigModel());
+            }
         } else if (clusterSyncDataModel.getType() == -1) {
             RegisterProviderModel registerProviderModel = clusterSyncDataModel.getRegisterProviderModel();
             RegisterCenterManager.down(registerProviderModel);
@@ -31,7 +39,9 @@ public class SyncClusterInformationHandler extends SimpleChannelInboundHandler<C
             RegisterProviderModel registerProviderModel = clusterSyncDataModel.getRegisterProviderModel();
             RegisterCenterManager.register(registerProviderModel);
         } else if (clusterSyncDataModel.getType() == 2) {
-            configCenterManager.save(clusterSyncDataModel.getProxyConfigModel(), clusterSyncDataModel.getConfigJson());
+            if (!expireHashMap.exists(clusterSyncDataModel.getSequenceId())) {
+                configCenterManager.save(clusterSyncDataModel.getProxyConfigModel(), clusterSyncDataModel.getConfigJson());
+            }
         }
     }
 }
